Extract navigation menu
